Revert "代码清理并更新版本为1.0.0.1"
This reverts commit 1ffceeb7d9cbc6c137588c87e9d4d35642781ec9.

+++ MainForm.cs
@@ -1,8 +1,15 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel;
+using System.Data;
 using System.Diagnostics;
 using System.Drawing;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices;
 using System.Security.Principal;
+using System.Text;
+using System.Threading.Tasks;
 using System.Windows.Forms;
 using Microsoft.Win32;
 using static EyeCarer.Properties.Resources;
@@ -15,7 +22,7 @@ namespace EyeCarer
         {
             InitializeComponent();
         }
-
+        
         private void MainForm_Load(object sender, EventArgs e)
         {
             #region 主窗口的Rectangle
@@ -301,7 +308,7 @@ namespace EyeCarer
             #region 计时器
             void Timer_Tick(object sender1, EventArgs e1)
             {
-                int usedSeconds = UsedTime.CountUsedTime(out bool isUsed) * SystemUsage.INTERVAL_SECONDS;
+                int usedSeconds = (int)(UsedTime.CountUsedTime(out bool isUsed) * SystemUsage.INTERVAL_SECONDS);
                 string usage = isUsed ? "正在使用" : "未使用";
                 if (WindowState == FormWindowState.Minimized)
                 {
